+++ jeap-reaction-observer-web/src/main/java/ch/admin/bit/jeap/reaction/observer/web/models/graph/EdgeDtoType.java
@@ -1,6 +1,6 @@
 package ch.admin.bit.jeap.reaction.observer.web.models.graph;
 
-public enum EdgeType {
+public enum EdgeDtoType {
     TRIGGER,
     ACTION
 }