+++ jeap-reaction-observer-web/src/main/java/ch/admin/bit/jeap/reaction/observer/web/models/graph/MessageNodeDto.java
@@ -20,4 +20,9 @@ public record MessageNodeDto(
     public long getId() {
         return id;
     }
+
+    @Override
+    public String getCanonicalId() {
+        return NodeDtoType.MESSAGE + "-" + id;
+    }
 }