+++ jeap-reaction-observer-web/src/main/java/ch/admin/bit/jeap/reaction/observer/web/models/graph/ReactionNodeDto.java
@@ -1,7 +1,6 @@
 package ch.admin.bit.jeap.reaction.observer.web.models.graph;
 
 import ch.admin.bit.jeap.reaction.observer.domain.models.graph.Reaction;
-import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record ReactionNodeDto(
         long id,
@@ -17,5 +16,10 @@ public record ReactionNodeDto(
     @Override
     public long getId() {
         return id;
+    }
+
+    @Override
+    public String getCanonicalId() {
+        return NodeDtoType.REACTION + "-" + id;
     }
 }