+++ jeap-reaction-observer-web/src/main/java/ch/admin/bit/jeap/reaction/observer/web/models/graph/TriggerEdgeDto.java
@@ -18,4 +18,9 @@ public record TriggerEdgeDto(
                 trigger.median()
         );
     }
+
+    @Override
+    public String getCanonicalId() {
+        return EdgeDtoType.TRIGGER + "-" + sourceId + "-" + targetReactionId;
+    }
 }