+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.23.0] - 2025-09-18
+
+### Added
+
+- Added new endpoint /api/graphs/components/{component-name} to retrieve component related reaction graph
+- Added new endpoint /api/graphs/messages/{message-type} to retrieve message related reaction graph
+
 ## [1.22.0] - 2025-09-17
 
 ### Added